requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
openai
faiss-

huggingface-hub

InstructorEmbedding
sentence-transformers